# models/create_product_not_implemented_body.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// CreateProductNotImplementedBody CreateProductNotImplementedBody create product not implemented body
//
// swagger:model CreateProductNotImplementedBody
type CreateProductNotImplementedBody struct {

	// message
	Message string `json:"message,omitempty"`
}

// Validate validates this create product not implemented body
func (m *CreateProductNotImplementedBody) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this create product not implemented body based on context it is used
func (m *CreateProductNotImplementedBody) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *CreateProductNotImplementedBody) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *CreateProductNotImplementedBody) UnmarshalBinary(b []byte) error {
	var res CreateProductNotImplementedBody
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}