# models/create_product_unprocessable_entity_body.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// CreateProductUnprocessableEntityBody CreateProductUnprocessableEntityBody create product unprocessable entity body
//
// swagger:model CreateProductUnprocessableEntityBody
type CreateProductUnprocessableEntityBody struct {

	// messages
	Messages []string `json:"messages"`
}

// Validate validates this create product unprocessable entity body
func (m *CreateProductUnprocessableEntityBody) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this create product unprocessable entity body based on context it is used
func (m *CreateProductUnprocessableEntityBody) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *CreateProductUnprocessableEntityBody) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *CreateProductUnprocessableEntityBody) UnmarshalBinary(b []byte) error {
	var res CreateProductUnprocessableEntityBody
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}