# models/update_product_unprocessable_entity_body.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// UpdateProductUnprocessableEntityBody UpdateProductUnprocessableEntityBody update product unprocessable entity body
//
// swagger:model UpdateProductUnprocessableEntityBody
type UpdateProductUnprocessableEntityBody struct {

	// messages
	Messages []string `json:"messages"`
}

// Validate validates this update product unprocessable entity body
func (m *UpdateProductUnprocessableEntityBody) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this update product unprocessable entity body based on context it is used
func (m *UpdateProductUnprocessableEntityBody) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *UpdateProductUnprocessableEntityBody) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *UpdateProductUnprocessableEntityBody) UnmarshalBinary(b []byte) error {
	var res UpdateProductUnprocessableEntityBody
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}